separate component for gene search results holder and states modify by passing in
