+++ src/components/initial_search_pages/gene_search_folder/GeneSearchPage.js
@@ -1,75 +1,20 @@
 import React, { useState, useEffect } from "react";
 import TextField from "@mui/material/TextField";
 import "./GeneSearchPage.css";
-import SampleList from "./SampleList";
+
 import SliderGene from "./SliderGene";
-import axios from "axios";
-import { useAuth0 } from "@auth0/auth0-react";
 
 import DashboardSidebar from "../../dashboard_side/dashboardSidebar";
+import GeneSearchResultsHolder from "./GeneSearchResultsHolder";
 
 import "../../bootstrap_gene_page/vendor/fontawesome-free/css/all.min.css";
 import "../../bootstrap_gene_page/css/sb-admin-2.min.css";
 
-import { clone } from "ramda";
-
 function GeneSearchPage() {
-  const [searchResult, setSearchResult] = useState([]);
+  
   //   Space so that user can run "blank" search
   const [searchInput, setSearchInput] = useState(" ");
-  const [listPage, setListPage] = useState(1);
-  const [isMounted, setIsMounted] = useState(false);
-  const [hasSearched, setHasSearched] = useState(false);
-
-  const { user } = useAuth0();
-
-  useEffect(() => {
-    if (isMounted) {
-      //handleSearch();
-    } else {
-      //setIsMounted(true);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [listPage]);
-
-  //   Url to search gene by keywords: 'api/gene/search/<str:search_word>/<str:page_id>'
-  const handleSearch = async () => {
-    try {
-      let search_input_for_url = clone(searchInput)
-
-      if (search_input_for_url === "") {
-        search_input_for_url = " ";
-      }
-      const response = await axios.get(
-        `${process.env.REACT_APP_BACKEND_URL
-        }/api/gene_search/${search_input_for_url}/${listPage.toString()}`
-      );
-      setSearchResult(response.data.genes);
-
-      console.log("search gene results: ")
-      console.log(response.data)
-
-      setHasSearched(true);
-    } catch (error) {
-      console.error(error);
-    }
-    console.log(listPage)
-  };
-
-  const handleIncrementPage = async () => {
-    console.log(user)
-    setListPage(listPage + 1);
-    handleSearch();
-    // console.log(listPage)
-  }
-
-  const handleDecrementPage = () => {
-    if (listPage > 1) {
-      setListPage(listPage - 1);
-      handleSearch();
-    }
-    // console.log(listPage)
-  }
+  const [clickedSearch, setClickedSearch] = useState(false);
 
   return (
     <body id="page-top">
@@ -99,7 +44,7 @@ function GeneSearchPage() {
                     label="Search by gene names or dataset name"
                   />
 
-                  <button type="submit" onClick={handleSearch} class="btn btn-primary" id="search_gene_button" aria-label="search">
+                  <button type="submit" onClick={ (e) => { setClickedSearch(true) }} class="btn btn-primary" id="search_gene_button" aria-label="search">
                     <i class="fas fa-search"></i>
                   </button>
                 </div>
@@ -107,32 +52,7 @@ function GeneSearchPage() {
                 <div id="gene_search_results_display_container">
                   <div class="card shadow" id="gene_search_results_display">
 
-                    <SampleList resultList={searchResult} />
-                    {!hasSearched && (
-                      <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', minWidth: '100%', maxWidth: '100%' }}>
-                        Start Searching!
-                      </div>
-                    )}
-                    {hasSearched && searchResult.length === 0 && (
-                      <div>
-                        <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', minWidth: '100%', maxWidth: '100%' }}>
-                          No results
-                        </div>
-                        {listPage > 1 ? <div className="float-left">
-                          <button onClick={handleDecrementPage}>Prev Page</button>
-                        </div> : <div></div>}
-                      </div>
-                    )}
-                    {searchResult.length > 0 && (
-                      <div>
-                        {listPage > 1 ? <div className="float-left">
-                          <button onClick={handleDecrementPage}>Prev Page</button>
-                        </div> : <div></div>}
-                        <div className="float-right">
-                          <button onClick={handleIncrementPage}>Next Page</button>
-                        </div>
-                      </div>
-                    )}
+                    <GeneSearchResultsHolder input_search_keyword={searchInput} performSearch={clickedSearch} setPerformSearch={setClickedSearch} />
 
                   </div>
                 </div>
